Rename pacckage from view_model to viewModel
