+++ app/src/main/java/ru/practicum/android/diploma/di/ViewModelModule.kt
@@ -2,7 +2,7 @@ package ru.practicum.android.diploma.di
 
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
-import ru.practicum.android.diploma.presentation.view_model.JobSearchViewModel
+import ru.practicum.android.diploma.presentation.viewModel.JobSearchViewModel
 
 val viewModelModule = module {
